add check_after_boot.py check_after_baxter.py check_after_setup.py check_state.py
